Improve interface and result

+++ lib/Koha/Contrib/ARK/Update.pm
@@ -1,35 +1,28 @@
-package Koha::Contrib::ARK::Updater;
+package Koha::Contrib::ARK::Update;
 # ABSTRACT: Update Koha ARK fields
+
+
 use Moose;
 
-with 'AnyEvent::Processor::Converter';
+with 'Koha::Contrib::ARK::Action';
 
 use Modern::Perl;
-use JSON;
-use YAML;
-use C4::Context;
-use C4::Biblio;
-use Try::Tiny;
-
-=attr ark
-
-L<Koha::Contrib::ARK> object.
-
-=cut
-has ark => ( is => 'rw', isa => 'Koha::Contrib::ARK' );
 
 
-sub convert {
-    my ($self, $br) = @_;
-    my ($biblionumber, $record) = @$br;
+
+sub action {
+    my ($self, $biblionumber, $record) = @_;
 
     my $a = $self->ark->c->{ark};
     my $ark = $self->ark->build_ark($biblionumber, $record);
-    $self->ark->log->info("Generated ARK: $ark\n");
     my $kfield = $a->{koha}->{ark};
     if ( $kfield->{letter} ) { # datafield
         if ( my $field = $record->field($kfield->{tag}) ) {
-            my @subf = grep { $_->[0] ne $kfield->{letter}; } @{$field->subf};
+            my @subf = grep {
+                my $keep = $_->[0] ne $kfield->{letter};
+                $self->ark->what_append('remove_existing') unless $keep;
+                $keep;
+            } @{$field->subf};
             push @subf, [ $kfield->{letter} => $ark ];
             $field->subf( \@subf );
         }
@@ -39,13 +32,15 @@ sub convert {
         }
     }
     else {
-        $record->delete($kfield->{tag});
+        if ( $record->field($kfield->{tag}) ) {
+            $record->delete($kfield->{tag});
+            $self->ark->what_append('remove_existing');
+        }
         $record->append( MARC::Moose::Field::Control->new(
             tag => $kfield->{tag},
             value => $ark ) );
     }
-
-    return [$biblionumber, $record];
+    $self->ark->what_append('add');
 }
